LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v3/default-src/manifest-v3-default-src-block-wildcard-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/extensions/resources/script.js because it appears in neither the script-src directive nor the default-src directive of the Content Security Policy.

Test that CSP for manifest V3 extensions does not allow wildcards for the default-src directive.

PASS
